+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
  - Fixed a potential `GL_INVALID_ENUM` error generated at initialization.
+ - Fixed `get_internal_format_if_supported()` sometimes panicking.
 
 ## Version 0.5.3