+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Fixed a potential `GL_INVALID_ENUM` error generated at initialization.
  - Fixed `get_internal_format_if_supported()` sometimes panicking.
+ - Fixed a panic with a tessellation evaluation shader that outputs quads.
 
 ## Version 0.5.3